Refactor mailchimp synchronization's domain object.
