+++ app/controllers/mailchimp_synchronizations_controller.rb
@@ -5,7 +5,7 @@ class MailchimpSynchronizationsController < ApplicationController
   def create
     respond_to do |format|
       format.js do
-        MailchimpSynchronizationJob.new(permitted_params[:group_id]).enqueue!
+        MailchimpSynchronizationJob.new(permitted_params[:mailing_list_id]).enqueue!
         flash[:notice] = translate(:mailchimp_synchronization_enqueued, email: current_person.email)
         #TODO: Properly localize flash message's text
         redirect_to(action: :index, controller: :subscriptions)
